Rename "accepted_conn" to "accepted_connections_total"
The authors understanding of the intended consumption of metrics is to
be able to review them in the Prometheus GUI, and understand what
queries to write to get an understanding of the health of the system. To
this end, certain guidelines have been formed:

- Use a single unit, and
- Where the metric is a count, suffix it with "total"

This makes it clear during debugging that the consumer should use things
such as the "rate" query on this metric to determine rate, rather than
needing to guess it's intended purpose.

This further renames "conn" to "connections", under the guidance that a
single base unit should be used, and that "connections" is more likely
than "conn" to be both understood and used across software packages.