Rename "listen_queue" to "listen_queue_connections"
Previous work began modifying the metrics such that they better matched
the Prometheus "best practice" guidelines as documented at the following
URL:

  https://prometheus.io/docs/practices/naming/

That work further justifies why such changes are made. This comment
extends that logic to the "listen_queue" metric.

This metric violated the "suffix with a consistent unit". The unit in
this case is "connections", and has been modified accordingly.